+++ src/main/java/de/maxhenkel/peek/data/DataStore.java
@@ -1,0 +1,12 @@
+package de.maxhenkel.peek.data;
+
+import net.minecraft.world.level.block.entity.ShulkerBoxBlockEntity;
+
+import javax.annotation.Nullable;
+
+public class DataStore {
+
+    @Nullable
+    public static ShulkerBoxBlockEntity lastOpenedShulkerBox;
+
+}